+++ R/psislw.R
@@ -14,12 +14,10 @@
 #' @param wtrunc For truncating very large weights to \eqn{S}^`wtrunc`. Set
 #'   to zero for no truncation.
 #' @param cores The number of cores to use for parallelization. This defaults to
-#'   the option `mc.cores` which can be set for an entire R session by
-#'   `options(mc.cores = NUMBER)`, the old option `loo.cores` is now
-#'   deprecated but will be given precedence over `mc.cores` until it is
-#'   removed. **As of version 2.0.0, the default is now 1 core if
-#'   `mc.cores` is not set, but we recommend using as many (or close to as
-#'   many) cores as possible.**
+#'   the option `mc.cores`, which can be set for an entire R session by
+#'   `options(mc.cores = NUMBER)`. **As of version 2.0.0, the default is now 1
+#'   core if `mc.cores` is not set, but we recommend using as many (or close to
+#'   as many) cores as possible.**
 #' @param llfun,llargs See [loo.function()].
 #' @param ... Ignored when `psislw()` is called directly. The `...` is
 #'   only used internally when `psislw()` is called by the [loo()]
@@ -40,7 +38,7 @@ psislw <- function(lw, wcp = 0.2, wtrunc = 3/4,
                    ...) {
   .Deprecated("psis")
 
-  cores <- loo_cores(cores)
+  cores <- loo_cores(cores, call = match.call())
 
   .psis <- function(lw_i) {
     x <- lw_i - max(lw_i)