+++ tests/testthat/test_psis.R
@@ -1,5 +1,4 @@
-options(mc.cores = 1)
-options(loo.cores = NULL)
+options(mc.cores = NULL)
 set.seed(123)
 
 LLarr <- example_loglik_array()