+++ tests/testthat/test_relative_eff.R
@@ -1,4 +1,4 @@
-options(mc.cores = 1)
+options(mc.cores = NULL)
 set.seed(123)
 
 LLarr <- example_loglik_array()